requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
frozendict==2.4.6
frozenlist==1.8.0
idna==3.11
joblib==1.5.2
multidict==6.7.0
multitasking==0.0.12
numpy==2.3.4
pandas==2.3.3
peewee==3.18.2
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytimeparse==1.1.8
pytz==2025.2
requests==2.32.5
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
soupsieve==2.8
ta==0.10.2
threadpoolctl==3.6.0
twelvedata==1.2.25
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
